cheatsheet: nest parser inside a toplevel cheatsheet pkg
